requirements.txt
asgiref==3.4.1
boto3==1.17.105
botocore==1.20.105
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Cython==0.29.23
fastapi==0.65.2
fuzzywuzzy==0.18.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
jmespath==0.10.0
numpy==1.21.0
pandas==1.2.5
priority==2.0.0
pydantic==1.8.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wsproto==1.0.0
